chore: finish Next public surface P0 placement
